refactor(daemon): split feishu cli final slice
