+++ crates/daemon/src/feishu_cli.rs
@@ -21,6 +21,17 @@ use crate::feishu_support::{
 
 const DEFAULT_FEISHU_REDIRECT_URI: &str = "http://127.0.0.1:34819/callback";
 
+#[path = "feishu_cli/bitable.rs"]
+mod bitable;
+#[path = "feishu_cli/render.rs"]
+mod render;
+#[cfg(test)]
+#[path = "feishu_cli/render_tests.rs"]
+mod render_tests;
+
+pub use self::bitable::*;
+use self::render::*;
+
 fn active_cli_command_name() -> &'static str {
     mvp::config::active_cli_command_name()
 }
@@ -2040,814 +2051,6 @@ pub async fn execute_feishu_calendar_freebusy(
     }))
 }
 
-pub async fn execute_feishu_bitable_list_tables(
-    args: &FeishuBitableListTablesArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["base:table:read"],
-        "loong feishu bitable list-tables",
-    )?;
-    let client = context.build_client()?;
-    let result = mvp::channel::feishu::api::resources::bitable::list_bitable_tables(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        args.page_token.as_deref(),
-        args.page_size,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "tables": result.items,
-        "has_more": result.has_more,
-        "page_token": result.page_token,
-    }))
-}
-
-pub async fn execute_feishu_bitable_app_create(
-    args: &FeishuBitableAppCreateArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable app-create",
-    )?;
-    let client = context.build_client()?;
-    let app = mvp::channel::feishu::api::resources::bitable::create_bitable_app(
-        &client,
-        &grant.access_token,
-        &args.name,
-        args.folder_token.as_deref(),
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "app": app,
-    }))
-}
-
-pub async fn execute_feishu_bitable_app_get(args: &FeishuBitableAppGetArgs) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable app-get",
-    )?;
-    let client = context.build_client()?;
-    let app = mvp::channel::feishu::api::resources::bitable::get_bitable_app(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "app": app,
-    }))
-}
-
-pub async fn execute_feishu_bitable_app_list(args: &FeishuBitableAppListArgs) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["drive:drive:readonly"],
-        "loong feishu bitable app-list",
-    )?;
-    let client = context.build_client()?;
-    let result = mvp::channel::feishu::api::resources::bitable::list_bitable_apps(
-        &client,
-        &grant.access_token,
-        &mvp::channel::feishu::api::resources::bitable::BitableAppListQuery {
-            folder_token: args.folder_token.clone(),
-            page_size: args.page_size,
-            page_token: args.page_token.clone(),
-        },
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "apps": result.apps,
-        "page_token": result.page_token,
-        "has_more": result.has_more,
-    }))
-}
-
-pub async fn execute_feishu_bitable_app_patch(
-    args: &FeishuBitableAppPatchArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable app-patch",
-    )?;
-    let client = context.build_client()?;
-    let app = mvp::channel::feishu::api::resources::bitable::patch_bitable_app(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        args.name.as_deref(),
-        args.is_advanced,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "app": app,
-    }))
-}
-
-pub async fn execute_feishu_bitable_app_copy(args: &FeishuBitableAppCopyArgs) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable app-copy",
-    )?;
-    let client = context.build_client()?;
-    let app = mvp::channel::feishu::api::resources::bitable::copy_bitable_app(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.name,
-        args.folder_token.as_deref(),
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "app": app,
-    }))
-}
-
-pub async fn execute_feishu_bitable_create_record(
-    args: &FeishuBitableCreateRecordArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["base:record:create"],
-        "loong feishu bitable create-record",
-    )?;
-    let client = context.build_client()?;
-    let fields = serde_json::from_str::<Value>(&args.fields)
-        .map_err(|error| format!("invalid --fields JSON: {error}"))?;
-    if !fields.is_object() {
-        return Err("--fields must be a JSON object (e.g. '{\"Name\": \"value\"}')".to_owned());
-    }
-    let record = mvp::channel::feishu::api::resources::bitable::create_bitable_record(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        fields,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "record": record,
-    }))
-}
-
-pub async fn execute_feishu_bitable_create_table(
-    args: &FeishuBitableCreateTableArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable create-table",
-    )?;
-    let client = context.build_client()?;
-    let fields = args
-        .fields
-        .as_deref()
-        .map(serde_json::from_str::<Value>)
-        .transpose()
-        .map_err(|error| format!("invalid --fields JSON: {error}"))?;
-    let fields = match fields {
-        Some(Value::Array(items)) => Some(items),
-        Some(_) => return Err("--fields must be a JSON array".to_owned()),
-        None => None,
-    };
-    let result = mvp::channel::feishu::api::resources::bitable::create_bitable_table(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.name,
-        args.default_view_name.as_deref(),
-        fields,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "result": result,
-    }))
-}
-
-pub async fn execute_feishu_bitable_patch_table(
-    args: &FeishuBitablePatchTableArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable patch-table",
-    )?;
-    let client = context.build_client()?;
-    let result = mvp::channel::feishu::api::resources::bitable::patch_bitable_table(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        &args.name,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "result": result,
-    }))
-}
-
-pub async fn execute_feishu_bitable_batch_create_tables(
-    args: &FeishuBitableBatchCreateTablesArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable batch-create-tables",
-    )?;
-    let client = context.build_client()?;
-    let tables = serde_json::from_str::<Value>(&args.tables)
-        .map_err(|error| format!("invalid --tables JSON: {error}"))?;
-    let tables = match tables {
-        Value::Array(items) => items,
-        Value::Null | Value::Bool(_) | Value::Number(_) | Value::String(_) | Value::Object(_) => {
-            return Err("--tables must be a JSON array".to_owned());
-        }
-    };
-    let result = mvp::channel::feishu::api::resources::bitable::batch_create_bitable_tables(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        tables,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "result": result,
-    }))
-}
-
-pub async fn execute_feishu_bitable_search_records(
-    args: &FeishuBitableSearchRecordsArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["base:record:retrieve"],
-        "loong feishu bitable search-records",
-    )?;
-    let client = context.build_client()?;
-    let filter = args
-        .filter
-        .as_deref()
-        .map(serde_json::from_str::<Value>)
-        .transpose()
-        .map_err(|error| format!("invalid --filter JSON: {error}"))?;
-    let sort = args
-        .sort
-        .as_deref()
-        .map(serde_json::from_str::<Value>)
-        .transpose()
-        .map_err(|error| format!("invalid --sort JSON: {error}"))?;
-    let result = mvp::channel::feishu::api::resources::bitable::search_bitable_records(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        &mvp::channel::feishu::api::resources::bitable::BitableRecordSearchQuery {
-            page_token: args.page_token.clone(),
-            page_size: args.page_size,
-            view_id: args.view_id.clone(),
-            filter,
-            sort,
-            field_names: (!args.field_names.is_empty()).then(|| args.field_names.clone()),
-            automatic_fields: args.automatic_fields.then_some(true),
-        },
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "result": result,
-    }))
-}
-
-pub async fn execute_feishu_bitable_update_record(
-    args: &FeishuBitableUpdateRecordArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["base:record:write"],
-        "loong feishu bitable update-record",
-    )?;
-    let client = context.build_client()?;
-    let fields = serde_json::from_str::<Value>(&args.fields)
-        .map_err(|error| format!("invalid --fields JSON: {error}"))?;
-    if !fields.is_object() {
-        return Err("--fields must be a JSON object (e.g. '{\"Name\": \"value\"}')".to_owned());
-    }
-    let record = mvp::channel::feishu::api::resources::bitable::update_bitable_record(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        &args.record_id,
-        fields,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "record": record,
-    }))
-}
-
-pub async fn execute_feishu_bitable_delete_record(
-    args: &FeishuBitableDeleteRecordArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["base:record:write"],
-        "loong feishu bitable delete-record",
-    )?;
-    let client = context.build_client()?;
-    let result = mvp::channel::feishu::api::resources::bitable::delete_bitable_record(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        &args.record_id,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "deleted": result.deleted,
-        "record_id": result.record_id,
-    }))
-}
-
-pub async fn execute_feishu_bitable_batch_create_records(
-    args: &FeishuBitableBatchCreateRecordsArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["base:record:write"],
-        "loong feishu bitable batch-create-records",
-    )?;
-    let client = context.build_client()?;
-    let records = serde_json::from_str::<Value>(&args.records)
-        .map_err(|error| format!("invalid --records JSON: {error}"))?;
-    let records = match records {
-        Value::Array(items) => items,
-        Value::Null | Value::Bool(_) | Value::Number(_) | Value::String(_) | Value::Object(_) => {
-            return Err("--records must be a JSON array".to_owned());
-        }
-    };
-    if records.len() > 500 {
-        return Err(format!(
-            "feishu.bitable.record.batch_create: batch size must be <= 500, got {}",
-            records.len()
-        ));
-    }
-    let result = mvp::channel::feishu::api::resources::bitable::batch_create_bitable_records(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        records,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "result": result,
-    }))
-}
-
-pub async fn execute_feishu_bitable_batch_update_records(
-    args: &FeishuBitableBatchUpdateRecordsArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["base:record:write"],
-        "loong feishu bitable batch-update-records",
-    )?;
-    let client = context.build_client()?;
-    let records = serde_json::from_str::<Value>(&args.records)
-        .map_err(|error| format!("invalid --records JSON: {error}"))?;
-    let records = match records {
-        Value::Array(items) => items,
-        Value::Null | Value::Bool(_) | Value::Number(_) | Value::String(_) | Value::Object(_) => {
-            return Err("--records must be a JSON array".to_owned());
-        }
-    };
-    if records.len() > 500 {
-        return Err(format!(
-            "feishu.bitable.record.batch_update: batch size must be <= 500, got {}",
-            records.len()
-        ));
-    }
-    let result = mvp::channel::feishu::api::resources::bitable::batch_update_bitable_records(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        records,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "result": result,
-    }))
-}
-
-pub async fn execute_feishu_bitable_batch_delete_records(
-    args: &FeishuBitableBatchDeleteRecordsArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["base:record:write"],
-        "loong feishu bitable batch-delete-records",
-    )?;
-    let client = context.build_client()?;
-    let records = serde_json::from_str::<Value>(&args.records)
-        .map_err(|error| format!("invalid --records JSON: {error}"))?;
-    let records = match records {
-        Value::Array(items) => items
-            .into_iter()
-            .map(|item| {
-                item.as_str()
-                    .map(ToOwned::to_owned)
-                    .ok_or_else(|| "--records must be a JSON array of strings".to_owned())
-            })
-            .collect::<CliResult<Vec<_>>>()?,
-        Value::Null | Value::Bool(_) | Value::Number(_) | Value::String(_) | Value::Object(_) => {
-            return Err("--records must be a JSON array".to_owned());
-        }
-    };
-    if records.len() > 500 {
-        return Err(format!(
-            "feishu.bitable.record.batch_delete: batch size must be <= 500, got {}",
-            records.len()
-        ));
-    }
-    let result = mvp::channel::feishu::api::resources::bitable::batch_delete_bitable_records(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        records,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "result": result,
-    }))
-}
-
-pub async fn execute_feishu_bitable_create_field(
-    args: &FeishuBitableCreateFieldArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable create-field",
-    )?;
-    let client = context.build_client()?;
-    let property = args
-        .property
-        .as_deref()
-        .map(serde_json::from_str::<Value>)
-        .transpose()
-        .map_err(|error| format!("invalid --property JSON: {error}"))?;
-    let field = mvp::channel::feishu::api::resources::bitable::create_bitable_field(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        &args.field_name,
-        args.field_type,
-        property,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "field": field,
-    }))
-}
-
-pub async fn execute_feishu_bitable_list_fields(
-    args: &FeishuBitableListFieldsArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable list-fields",
-    )?;
-    let client = context.build_client()?;
-    let result = mvp::channel::feishu::api::resources::bitable::list_bitable_fields(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        &mvp::channel::feishu::api::resources::bitable::BitableFieldListQuery {
-            view_id: args.view_id.clone(),
-            page_size: args.page_size,
-            page_token: args.page_token.clone(),
-        },
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "fields": result.items,
-        "page_token": result.page_token,
-        "has_more": result.has_more,
-        "total": result.total,
-    }))
-}
-
-pub async fn execute_feishu_bitable_update_field(
-    args: &FeishuBitableUpdateFieldArgs,
-) -> CliResult<Value> {
-    let field_name = args
-        .field_name
-        .as_deref()
-        .ok_or_else(|| "--field-name and --type are required for field update".to_owned())?;
-    let field_type = args
-        .field_type
-        .ok_or_else(|| "--field-name and --type are required for field update".to_owned())?;
-
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable update-field",
-    )?;
-    let client = context.build_client()?;
-    let property = args
-        .property
-        .as_deref()
-        .map(serde_json::from_str::<Value>)
-        .transpose()
-        .map_err(|error| format!("invalid --property JSON: {error}"))?;
-    let field = mvp::channel::feishu::api::resources::bitable::update_bitable_field(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        &args.field_id,
-        field_name,
-        field_type,
-        property,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "field": field,
-    }))
-}
-
-pub async fn execute_feishu_bitable_delete_field(
-    args: &FeishuBitableDeleteFieldArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable delete-field",
-    )?;
-    let client = context.build_client()?;
-    let result = mvp::channel::feishu::api::resources::bitable::delete_bitable_field(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        &args.field_id,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "deleted": result.deleted,
-        "field_id": result.field_id,
-    }))
-}
-
-pub async fn execute_feishu_bitable_create_view(
-    args: &FeishuBitableCreateViewArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable create-view",
-    )?;
-    let client = context.build_client()?;
-    let view = mvp::channel::feishu::api::resources::bitable::create_bitable_view(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        &args.view_name,
-        args.view_type.as_deref(),
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "view": view,
-    }))
-}
-
-pub async fn execute_feishu_bitable_get_view(args: &FeishuBitableGetViewArgs) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable get-view",
-    )?;
-    let client = context.build_client()?;
-    let view = mvp::channel::feishu::api::resources::bitable::get_bitable_view(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        &args.view_id,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "view": view,
-    }))
-}
-
-pub async fn execute_feishu_bitable_list_views(
-    args: &FeishuBitableListViewsArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable list-views",
-    )?;
-    let client = context.build_client()?;
-    let result = mvp::channel::feishu::api::resources::bitable::list_bitable_views(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        &mvp::channel::feishu::api::resources::bitable::BitableViewListQuery {
-            page_size: args.page_size,
-            page_token: args.page_token.clone(),
-        },
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "views": result.items,
-        "page_token": result.page_token,
-        "has_more": result.has_more,
-        "total": result.total,
-    }))
-}
-
-pub async fn execute_feishu_bitable_patch_view(
-    args: &FeishuBitablePatchViewArgs,
-) -> CliResult<Value> {
-    let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
-    ensure_grant_has_any_scope(
-        &grant,
-        context.resolved.configured_account_id.as_str(),
-        &["bitable:app"],
-        "loong feishu bitable patch-view",
-    )?;
-    let client = context.build_client()?;
-    let view = mvp::channel::feishu::api::resources::bitable::patch_bitable_view(
-        &client,
-        &grant.access_token,
-        &args.app_token,
-        &args.table_id,
-        &args.view_id,
-        &args.view_name,
-    )
-    .await?;
-
-    Ok(json!({
-        "account_id": context.account_id(),
-        "configured_account": context.resolved.configured_account_label,
-        "principal": grant.principal,
-        "view": view,
-    }))
-}
-
 pub async fn execute_feishu_send(args: &FeishuSendArgs) -> CliResult<Value> {
     let (context, grant) = load_context_and_fresh_grant(&args.grant).await?;
     let action = format!("{} feishu send", active_cli_command_name());
@@ -3234,1918 +2437,6 @@ fn serialize_grant_summary(
             required_scopes,
         ),
     })
-}
-
-#[allow(clippy::print_stdout)]
-fn print_feishu_payload(
-    payload: &Value,
-    as_json: bool,
-    render_text: fn(&Value) -> CliResult<String>,
-) -> CliResult<()> {
-    if as_json {
-        let encoded = serde_json::to_string_pretty(payload)
-            .map_err(|error| format!("serialize feishu command output failed: {error}"))?;
-        println!("{encoded}");
-        return Ok(());
-    }
-    println!("{}", render_text(payload)?);
-    Ok(())
-}
-
-fn render_onboard_text(payload: &Value) -> CliResult<String> {
-    let mut lines = vec![
-        "feishu onboard".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!("config: {}", required_json_string(payload, "config")?),
-        format!(
-            "credential_source: {}",
-            required_json_string(payload, "credential_source")?
-        ),
-        format!("domain: {}", required_json_string(payload, "domain")?),
-        format!("mode: {}", required_json_string(payload, "mode")?),
-    ]);
-    if let Some(owner_open_id) = payload.get("owner_open_id").and_then(Value::as_str) {
-        lines.push(format!("owner_open_id: {owner_open_id}"));
-    }
-    if let Some(bot_name) = payload.get("bot_name").and_then(Value::as_str) {
-        lines.push(format!("bot_name: {bot_name}"));
-    }
-    if let Some(bot_open_id) = payload.get("bot_open_id").and_then(Value::as_str) {
-        lines.push(format!("bot_open_id: {bot_open_id}"));
-    }
-    if let Some(qr_url) = payload.get("qr_url").and_then(Value::as_str) {
-        lines.push(format!("qr_url: {qr_url}"));
-    }
-    lines.push(format!(
-        "serve_command: {}",
-        required_json_string(payload, "serve_command")?
-    ));
-    lines.push(format!(
-        "status_command: {}",
-        required_json_string(payload, "status_command")?
-    ));
-    if let Some(notes) = payload.get("notes").and_then(Value::as_array) {
-        for note in notes.iter().filter_map(Value::as_str) {
-            lines.push(format!("note: {note}"));
-        }
-    }
-    Ok(lines.join("\n"))
-}
-
-fn render_auth_start_text(payload: &Value) -> CliResult<String> {
-    let mut lines = vec![
-        "feishu auth start".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!("state: {}", required_json_string(payload, "state")?),
-        format!(
-            "redirect_uri: {}",
-            required_json_string(payload, "redirect_uri")?
-        ),
-        format!(
-            "authorize_url: {}",
-            required_json_string(payload, "authorize_url")?
-        ),
-        format!(
-            "capabilities: {}",
-            payload
-                .get("capabilities")
-                .and_then(Value::as_array)
-                .map(|values| {
-                    values
-                        .iter()
-                        .filter_map(Value::as_str)
-                        .collect::<Vec<_>>()
-                        .join(", ")
-                })
-                .filter(|value| !value.is_empty())
-                .unwrap_or_else(|| "-".to_owned())
-        ),
-        format!(
-            "scopes: {}",
-            payload
-                .get("scopes")
-                .and_then(Value::as_array)
-                .map(|values| {
-                    values
-                        .iter()
-                        .filter_map(Value::as_str)
-                        .collect::<Vec<_>>()
-                        .join(", ")
-                })
-                .unwrap_or_default()
-        ),
-        format!(
-            "sqlite_path: {}",
-            required_json_string(payload, "sqlite_path")?
-        ),
-    ]);
-    if let Some(status) = payload.get("status").and_then(Value::as_str) {
-        lines.push(format!("status: {status}"));
-    }
-    Ok(lines.join("\n"))
-}
-
-fn render_auth_exchange_text(payload: &Value) -> CliResult<String> {
-    let principal = payload
-        .get("principal")
-        .cloned()
-        .unwrap_or_else(|| json!({}));
-    let mut lines = vec![
-        "feishu auth exchange".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "open_id: {}",
-            principal
-                .get("open_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "name: {}",
-            principal.get("name").and_then(Value::as_str).unwrap_or("-")
-        ),
-        format!(
-            "scopes: {}",
-            payload
-                .get("granted_scopes")
-                .and_then(Value::as_array)
-                .map(|values| {
-                    values
-                        .iter()
-                        .filter_map(Value::as_str)
-                        .collect::<Vec<_>>()
-                        .join(", ")
-                })
-                .unwrap_or_default()
-        ),
-        format!(
-            "selected_open_id: {}",
-            payload
-                .get("selected_open_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "effective_open_id: {}",
-            payload
-                .get("effective_open_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_auth_list_text(payload: &Value) -> CliResult<String> {
-    let grants = payload
-        .get("grants")
-        .and_then(Value::as_array)
-        .ok_or_else(|| "feishu auth list payload missing grants".to_owned())?;
-    let mut lines = vec![
-        "feishu auth list".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-        format!("grant_count: {}", grants.len()),
-        format!(
-            "selected_open_id: {}",
-            payload
-                .get("selected_open_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.insert(2, format!("configured_account: {configured_account}"));
-    }
-    if let Some(effective_open_id) = payload.get("effective_open_id").and_then(Value::as_str) {
-        lines.push(format!("effective_open_id: {effective_open_id}"));
-    }
-    if let Some(select_hint) = payload
-        .get("recommendations")
-        .and_then(|value| value.get("select_command"))
-        .and_then(Value::as_str)
-    {
-        lines.push(format!("select_hint: {select_hint}"));
-    }
-    if let Some(stale_selected_open_id) = payload
-        .get("recommendations")
-        .and_then(|value| value.get("stale_selected_open_id"))
-        .and_then(Value::as_str)
-    {
-        lines.push(format!("stale_selected_open_id: {stale_selected_open_id}"));
-    }
-    if let Some(auth_start_hint) = payload
-        .get("recommendations")
-        .and_then(|value| value.get("auth_start_command"))
-        .and_then(Value::as_str)
-    {
-        lines.push(format!("auth_start_hint: {auth_start_hint}"));
-    }
-    for grant in grants {
-        lines.push(render_auth_grant_summary_line(grant));
-    }
-    Ok(lines.join("\n"))
-}
-
-fn render_auth_select_text(payload: &Value) -> CliResult<String> {
-    let mut lines = vec![
-        "feishu auth select".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.push(format!(
-        "selected_open_id: {}",
-        required_json_string(payload, "selected_open_id")?
-    ));
-    if let Some(grant) = payload.get("grant") {
-        lines.push(render_auth_grant_summary_line(grant));
-    }
-    Ok(lines.join("\n"))
-}
-
-fn render_auth_status_text(payload: &Value) -> CliResult<String> {
-    if payload
-        .get("status_scope")
-        .and_then(Value::as_str)
-        .is_some_and(|scope| scope == "account")
-    {
-        let grants = payload
-            .get("grants")
-            .and_then(Value::as_array)
-            .ok_or_else(|| "feishu auth status payload missing grants".to_owned())?;
-        let mut lines = vec![
-            "feishu auth status".to_owned(),
-            format!("account: {}", required_json_string(payload, "account_id")?),
-            format!("status_scope: account"),
-            format!("grant_count: {}", grants.len()),
-            format!(
-                "selected_open_id: {}",
-                payload
-                    .get("selected_open_id")
-                    .and_then(Value::as_str)
-                    .unwrap_or("-")
-            ),
-        ];
-        if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str)
-        {
-            lines.insert(2, format!("configured_account: {configured_account}"));
-        }
-        if let Some(effective_open_id) = payload.get("effective_open_id").and_then(Value::as_str) {
-            lines.push(format!("effective_open_id: {effective_open_id}"));
-        }
-        if let Some(select_hint) = payload
-            .get("recommendations")
-            .and_then(|value| value.get("select_command"))
-            .and_then(Value::as_str)
-        {
-            lines.push(format!("select_hint: {select_hint}"));
-        }
-        if let Some(stale_selected_open_id) = payload
-            .get("recommendations")
-            .and_then(|value| value.get("stale_selected_open_id"))
-            .and_then(Value::as_str)
-        {
-            lines.push(format!("stale_selected_open_id: {stale_selected_open_id}"));
-        }
-        for grant in grants {
-            lines.push(render_auth_grant_summary_line(grant));
-        }
-        return Ok(lines.join("\n"));
-    }
-
-    let status = payload
-        .get("status")
-        .ok_or_else(|| "feishu auth status payload missing status".to_owned())?;
-    let doc_write_status = payload.get("doc_write_status").unwrap_or(&Value::Null);
-    let message_write_status = payload.get("message_write_status").unwrap_or(&Value::Null);
-    let auth_start_hint = payload
-        .get("recommendations")
-        .and_then(|value| value.get("auth_start_command"))
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let select_hint = payload
-        .get("recommendations")
-        .and_then(|value| value.get("select_command"))
-        .and_then(Value::as_str);
-    let selected_open_id = payload
-        .get("selected_open_id")
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let effective_open_id = payload
-        .get("effective_open_id")
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let mut lines = vec![
-        "feishu auth status".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-        format!(
-            "has_grant: {}",
-            status
-                .get("has_grant")
-                .and_then(Value::as_bool)
-                .unwrap_or(false)
-        ),
-        format!(
-            "access_token_expired: {}",
-            status
-                .get("access_token_expired")
-                .and_then(Value::as_bool)
-                .unwrap_or(true)
-        ),
-        format!(
-            "refresh_token_expired: {}",
-            status
-                .get("refresh_token_expired")
-                .and_then(Value::as_bool)
-                .unwrap_or(true)
-        ),
-        format!(
-            "missing_scopes: {}",
-            status
-                .get("missing_scopes")
-                .and_then(Value::as_array)
-                .map(|values| {
-                    values
-                        .iter()
-                        .filter_map(Value::as_str)
-                        .collect::<Vec<_>>()
-                        .join(", ")
-                })
-                .unwrap_or_default()
-        ),
-        format!(
-            "doc_write_ready: {}",
-            doc_write_status
-                .get("ready")
-                .and_then(Value::as_bool)
-                .unwrap_or(false)
-        ),
-        format!(
-            "matched_doc_write_scopes: {}",
-            doc_write_status
-                .get("matched_scopes")
-                .and_then(Value::as_array)
-                .map(|values| {
-                    values
-                        .iter()
-                        .filter_map(Value::as_str)
-                        .collect::<Vec<_>>()
-                        .join(", ")
-                })
-                .filter(|value| !value.is_empty())
-                .unwrap_or_else(|| "-".to_owned())
-        ),
-        format!(
-            "message_write_ready: {}",
-            message_write_status
-                .get("ready")
-                .and_then(Value::as_bool)
-                .unwrap_or(false)
-        ),
-        format!(
-            "matched_write_scopes: {}",
-            message_write_status
-                .get("matched_scopes")
-                .and_then(Value::as_array)
-                .map(|values| {
-                    values
-                        .iter()
-                        .filter_map(Value::as_str)
-                        .collect::<Vec<_>>()
-                        .join(", ")
-                })
-                .filter(|value| !value.is_empty())
-                .unwrap_or_else(|| "-".to_owned())
-        ),
-        format!("auth_start_hint: {auth_start_hint}"),
-        format!("selected_open_id: {selected_open_id}"),
-        format!("effective_open_id: {effective_open_id}"),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.insert(2, format!("configured_account: {configured_account}"));
-    }
-    if let Some(requested_open_id) = payload.get("requested_open_id").and_then(Value::as_str) {
-        lines.push(format!("requested_open_id: {requested_open_id}"));
-    }
-    if let Some(select_hint) = select_hint {
-        lines.push(format!("select_hint: {select_hint}"));
-    }
-    if let Some(available_open_ids) = payload
-        .get("available_open_ids")
-        .and_then(Value::as_array)
-        .filter(|values| !values.is_empty())
-    {
-        lines.push(format!(
-            "available_open_ids: {}",
-            available_open_ids
-                .iter()
-                .filter_map(Value::as_str)
-                .collect::<Vec<_>>()
-                .join(", ")
-        ));
-    }
-    Ok(lines.join("\n"))
-}
-
-fn render_auth_grant_summary_line(grant: &Value) -> String {
-    let principal = grant
-        .get("principal")
-        .and_then(Value::as_object)
-        .cloned()
-        .unwrap_or_default();
-    let status = grant
-        .get("status")
-        .and_then(Value::as_object)
-        .cloned()
-        .unwrap_or_default();
-    let doc_write_status = grant
-        .get("doc_write_status")
-        .and_then(Value::as_object)
-        .cloned()
-        .unwrap_or_default();
-    let message_write_status = grant
-        .get("message_write_status")
-        .and_then(Value::as_object)
-        .cloned()
-        .unwrap_or_default();
-    let missing_scopes = status
-        .get("missing_scopes")
-        .and_then(Value::as_array)
-        .map(|values| {
-            values
-                .iter()
-                .filter_map(Value::as_str)
-                .collect::<Vec<_>>()
-                .join(", ")
-        })
-        .filter(|value| !value.is_empty())
-        .unwrap_or_else(|| "-".to_owned());
-    let matched_doc_write_scopes = doc_write_status
-        .get("matched_scopes")
-        .and_then(Value::as_array)
-        .map(|values| {
-            values
-                .iter()
-                .filter_map(Value::as_str)
-                .collect::<Vec<_>>()
-                .join(", ")
-        })
-        .filter(|value| !value.is_empty())
-        .unwrap_or_else(|| "-".to_owned());
-    let matched_write_scopes = message_write_status
-        .get("matched_scopes")
-        .and_then(Value::as_array)
-        .map(|values| {
-            values
-                .iter()
-                .filter_map(Value::as_str)
-                .collect::<Vec<_>>()
-                .join(", ")
-        })
-        .filter(|value| !value.is_empty())
-        .unwrap_or_else(|| "-".to_owned());
-    let auth_start_hint = grant
-        .get("recommendations")
-        .and_then(|value| value.get("auth_start_command"))
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-
-    format!(
-        "open_id: {} | selected: {} | effective_selected: {} | name: {} | access_expired: {} | refresh_expired: {} | missing_scopes: {} | doc_write_ready: {} | matched_doc_write_scopes: {} | message_write_ready: {} | matched_write_scopes: {} | auth_start_hint: {}",
-        principal
-            .get("open_id")
-            .and_then(Value::as_str)
-            .unwrap_or("-"),
-        grant
-            .get("selected")
-            .and_then(Value::as_bool)
-            .unwrap_or(false),
-        grant
-            .get("effective_selected")
-            .and_then(Value::as_bool)
-            .unwrap_or(false),
-        principal.get("name").and_then(Value::as_str).unwrap_or("-"),
-        status
-            .get("access_token_expired")
-            .and_then(Value::as_bool)
-            .unwrap_or(true),
-        status
-            .get("refresh_token_expired")
-            .and_then(Value::as_bool)
-            .unwrap_or(true),
-        missing_scopes,
-        doc_write_status
-            .get("ready")
-            .and_then(Value::as_bool)
-            .unwrap_or(false),
-        matched_doc_write_scopes,
-        message_write_status
-            .get("ready")
-            .and_then(Value::as_bool)
-            .unwrap_or(false),
-        matched_write_scopes,
-        auth_start_hint,
-    )
-}
-
-fn render_auth_revoke_text(payload: &Value) -> CliResult<String> {
-    let mut lines = vec![
-        "feishu auth revoke".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "open_id: {}",
-            payload
-                .get("open_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "deleted: {}",
-            payload
-                .get("deleted")
-                .and_then(Value::as_bool)
-                .unwrap_or(false)
-        ),
-    ]);
-    if let Some(grant_count) = payload.get("grant_count").and_then(Value::as_u64) {
-        lines.push(format!("grant_count: {grant_count}"));
-    }
-    if payload.get("selected_open_id").is_some() {
-        lines.push(format!(
-            "selected_open_id: {}",
-            payload
-                .get("selected_open_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ));
-    }
-    if payload.get("effective_open_id").is_some() {
-        lines.push(format!(
-            "effective_open_id: {}",
-            payload
-                .get("effective_open_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ));
-    }
-    if let Some(select_hint) = payload
-        .get("recommendations")
-        .and_then(|value| value.get("select_command"))
-        .and_then(Value::as_str)
-    {
-        lines.push(format!("select_hint: {select_hint}"));
-    }
-    if let Some(auth_start_hint) = payload
-        .get("recommendations")
-        .and_then(|value| value.get("auth_start_command"))
-        .and_then(Value::as_str)
-    {
-        lines.push(format!("auth_start_hint: {auth_start_hint}"));
-    }
-    Ok(lines.join("\n"))
-}
-
-fn render_whoami_text(payload: &Value) -> CliResult<String> {
-    let principal = payload
-        .get("principal")
-        .ok_or_else(|| "feishu whoami payload missing principal".to_owned())?;
-    let mut lines = vec![
-        "feishu whoami".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "open_id: {}",
-            principal
-                .get("open_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "name: {}",
-            principal.get("name").and_then(Value::as_str).unwrap_or("-")
-        ),
-        format!(
-            "email: {}",
-            principal
-                .get("email")
-                .and_then(Value::as_str)
-                .or_else(|| principal.get("enterprise_email").and_then(Value::as_str))
-                .unwrap_or("-")
-        ),
-        format!(
-            "tenant_key: {}",
-            principal
-                .get("tenant_key")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_read_doc_text(payload: &Value) -> CliResult<String> {
-    let document = payload
-        .get("document")
-        .ok_or_else(|| "feishu read doc payload missing document".to_owned())?;
-    let mut lines = vec![
-        "feishu read doc".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.push(format!(
-        "document_id: {}",
-        document
-            .get("document_id")
-            .and_then(Value::as_str)
-            .unwrap_or("-")
-    ));
-    lines.push(String::new());
-    lines.push(
-        document
-            .get("content")
-            .and_then(Value::as_str)
-            .unwrap_or_default()
-            .to_owned(),
-    );
-    Ok(lines.join("\n"))
-}
-
-fn render_doc_create_text(payload: &Value) -> CliResult<String> {
-    let document = payload
-        .get("document")
-        .ok_or_else(|| "feishu doc create payload missing document".to_owned())?;
-    let mut lines = vec![
-        "feishu doc create".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "document_id: {}",
-            document
-                .get("document_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "url: {}",
-            document.get("url").and_then(Value::as_str).unwrap_or("-")
-        ),
-        format!(
-            "content_inserted: {}",
-            payload
-                .get("content_inserted")
-                .and_then(Value::as_bool)
-                .unwrap_or(false)
-        ),
-        format!(
-            "inserted_block_count: {}",
-            payload
-                .get("inserted_block_count")
-                .and_then(Value::as_u64)
-                .unwrap_or_default()
-        ),
-        format!(
-            "insert_batch_count: {}",
-            payload
-                .get("insert_batch_count")
-                .and_then(Value::as_u64)
-                .unwrap_or_default()
-        ),
-        format!(
-            "content_type: {}",
-            payload
-                .get("content_type")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_doc_append_text(payload: &Value) -> CliResult<String> {
-    let document = payload
-        .get("document")
-        .ok_or_else(|| "feishu doc append payload missing document".to_owned())?;
-    let mut lines = vec![
-        "feishu doc append".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "document_id: {}",
-            document
-                .get("document_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "url: {}",
-            document.get("url").and_then(Value::as_str).unwrap_or("-")
-        ),
-        format!(
-            "inserted_block_count: {}",
-            payload
-                .get("inserted_block_count")
-                .and_then(Value::as_u64)
-                .unwrap_or_default()
-        ),
-        format!(
-            "insert_batch_count: {}",
-            payload
-                .get("insert_batch_count")
-                .and_then(Value::as_u64)
-                .unwrap_or_default()
-        ),
-        format!(
-            "content_type: {}",
-            payload
-                .get("content_type")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_messages_history_text(payload: &Value) -> CliResult<String> {
-    let page = payload
-        .get("page")
-        .ok_or_else(|| "feishu message history payload missing page".to_owned())?;
-    let mut lines = vec![
-        "feishu messages history".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "items: {}",
-            page.get("items")
-                .and_then(Value::as_array)
-                .map_or(0, std::vec::Vec::len)
-        ),
-        format!(
-            "has_more: {}",
-            page.get("has_more")
-                .and_then(Value::as_bool)
-                .unwrap_or(false)
-        ),
-        format!(
-            "page_token: {}",
-            page.get("page_token")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_messages_get_text(payload: &Value) -> CliResult<String> {
-    let message = payload
-        .get("message")
-        .ok_or_else(|| "feishu message get payload missing message".to_owned())?;
-    let encoded = serde_json::to_string_pretty(message)
-        .map_err(|error| format!("serialize feishu message detail failed: {error}"))?;
-    let mut lines = vec![
-        "feishu messages get".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.push(encoded);
-    Ok(lines.join("\n"))
-}
-
-fn render_messages_resource_text(payload: &Value) -> CliResult<String> {
-    let mut lines = vec![
-        "feishu messages resource".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "message_id: {}",
-            payload
-                .get("message_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "file_key: {}",
-            payload
-                .get("file_key")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "type: {}",
-            payload
-                .get("resource_type")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "path: {}",
-            payload.get("path").and_then(Value::as_str).unwrap_or("-")
-        ),
-        format!(
-            "bytes_written: {}",
-            payload
-                .get("bytes_written")
-                .and_then(Value::as_u64)
-                .unwrap_or_default()
-        ),
-        format!(
-            "content_type: {}",
-            payload
-                .get("content_type")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "file_name: {}",
-            payload
-                .get("file_name")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_send_text(payload: &Value) -> CliResult<String> {
-    let delivery = payload
-        .get("delivery")
-        .ok_or_else(|| "feishu send payload missing delivery".to_owned())?;
-    let mut lines = vec![
-        "feishu send".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "message_id: {}",
-            delivery
-                .get("message_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "receive_id_type: {}",
-            delivery
-                .get("receive_id_type")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "receive_id: {}",
-            delivery
-                .get("receive_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "uuid: {}",
-            delivery.get("uuid").and_then(Value::as_str).unwrap_or("-")
-        ),
-        format!(
-            "msg_type: {}",
-            delivery
-                .get("msg_type")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_reply_text(payload: &Value) -> CliResult<String> {
-    let delivery = payload
-        .get("delivery")
-        .ok_or_else(|| "feishu reply payload missing delivery".to_owned())?;
-    let mut lines = vec![
-        "feishu reply".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "message_id: {}",
-            delivery
-                .get("message_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "reply_to_message_id: {}",
-            delivery
-                .get("reply_to_message_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "reply_in_thread: {}",
-            delivery
-                .get("reply_in_thread")
-                .and_then(Value::as_bool)
-                .unwrap_or(false)
-        ),
-        format!(
-            "uuid: {}",
-            delivery.get("uuid").and_then(Value::as_str).unwrap_or("-")
-        ),
-        format!(
-            "msg_type: {}",
-            delivery
-                .get("msg_type")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_search_messages_text(payload: &Value) -> CliResult<String> {
-    let page = payload
-        .get("page")
-        .ok_or_else(|| "feishu message search payload missing page".to_owned())?;
-    let mut lines = vec![
-        "feishu search messages".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "items: {}",
-            page.get("items")
-                .and_then(Value::as_array)
-                .map_or(0, std::vec::Vec::len)
-        ),
-        format!(
-            "has_more: {}",
-            page.get("has_more")
-                .and_then(Value::as_bool)
-                .unwrap_or(false)
-        ),
-        format!(
-            "page_token: {}",
-            page.get("page_token")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_calendar_list_text(payload: &Value) -> CliResult<String> {
-    if payload
-        .get("primary")
-        .and_then(Value::as_bool)
-        .unwrap_or(false)
-    {
-        let calendars = payload
-            .get("calendars")
-            .and_then(|value| value.get("calendars"))
-            .and_then(Value::as_array)
-            .map_or(0, std::vec::Vec::len);
-        let mut lines = vec![
-            "feishu calendar list".to_owned(),
-            format!("account: {}", required_json_string(payload, "account_id")?),
-        ];
-        if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str)
-        {
-            lines.push(format!("configured_account: {configured_account}"));
-        }
-        lines.extend([
-            "primary: true".to_owned(),
-            format!("calendars: {calendars}"),
-        ]);
-        return Ok(lines.join("\n"));
-    }
-
-    let page = payload
-        .get("page")
-        .ok_or_else(|| "feishu calendar list payload missing page".to_owned())?;
-    let mut lines = vec![
-        "feishu calendar list".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        "primary: false".to_owned(),
-        format!(
-            "calendars: {}",
-            page.get("calendar_list")
-                .and_then(Value::as_array)
-                .map_or(0, std::vec::Vec::len)
-        ),
-        format!(
-            "sync_token: {}",
-            page.get("sync_token")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_calendar_freebusy_text(payload: &Value) -> CliResult<String> {
-    let result = payload
-        .get("result")
-        .ok_or_else(|| "feishu calendar freebusy payload missing result".to_owned())?;
-    let mut lines = vec![
-        "feishu calendar freebusy".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.push(format!(
-        "slots: {}",
-        result
-            .get("freebusy_list")
-            .and_then(Value::as_array)
-            .map_or(0, std::vec::Vec::len)
-    ));
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_list_tables_text(payload: &Value) -> CliResult<String> {
-    let mut lines = vec![
-        "feishu bitable list-tables".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "tables: {}",
-            payload
-                .get("tables")
-                .and_then(Value::as_array)
-                .map_or(0, std::vec::Vec::len)
-        ),
-        format!(
-            "has_more: {}",
-            payload
-                .get("has_more")
-                .and_then(Value::as_bool)
-                .unwrap_or(false)
-        ),
-        format!(
-            "page_token: {}",
-            payload
-                .get("page_token")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_app_text(payload: &Value) -> CliResult<String> {
-    let app = payload
-        .get("app")
-        .ok_or_else(|| "feishu bitable app payload missing app".to_owned())?;
-    let mut lines = vec![
-        "feishu bitable app".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "app_token: {}",
-            app.get("app_token").and_then(Value::as_str).unwrap_or("-")
-        ),
-        format!(
-            "name: {}",
-            app.get("name").and_then(Value::as_str).unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_app_list_text(payload: &Value) -> CliResult<String> {
-    let mut lines = vec![
-        "feishu bitable app-list".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "apps: {}",
-            payload
-                .get("apps")
-                .and_then(Value::as_array)
-                .map_or(0, std::vec::Vec::len)
-        ),
-        format!(
-            "has_more: {}",
-            payload
-                .get("has_more")
-                .and_then(Value::as_bool)
-                .unwrap_or(false)
-        ),
-        format!(
-            "page_token: {}",
-            payload
-                .get("page_token")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_create_record_text(payload: &Value) -> CliResult<String> {
-    let record = payload
-        .get("record")
-        .ok_or_else(|| "feishu bitable create payload missing record".to_owned())?;
-    let mut lines = vec![
-        "feishu bitable create-record".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "record_id: {}",
-            record
-                .get("record_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "fields: {}",
-            record
-                .get("fields")
-                .and_then(Value::as_object)
-                .map_or(0, serde_json::Map::len)
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_table_text(payload: &Value) -> CliResult<String> {
-    let result = payload
-        .get("result")
-        .ok_or_else(|| "feishu bitable table payload missing result".to_owned())?;
-    let mut lines = vec![
-        "feishu bitable table".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "table_id: {}",
-            result
-                .get("table_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "name: {}",
-            result.get("name").and_then(Value::as_str).unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_table_batch_create_text(payload: &Value) -> CliResult<String> {
-    let result = payload
-        .get("result")
-        .ok_or_else(|| "feishu bitable batch create tables payload missing result".to_owned())?;
-    let mut lines = vec![
-        "feishu bitable batch-create-tables".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.push(format!(
-        "table_ids: {}",
-        result
-            .get("table_ids")
-            .and_then(Value::as_array)
-            .map_or(0, std::vec::Vec::len)
-    ));
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_search_records_text(payload: &Value) -> CliResult<String> {
-    let result = payload
-        .get("result")
-        .ok_or_else(|| "feishu bitable search payload missing result".to_owned())?;
-    let mut lines = vec![
-        "feishu bitable search-records".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "records: {}",
-            result
-                .get("items")
-                .and_then(Value::as_array)
-                .map_or(0, std::vec::Vec::len)
-        ),
-        format!(
-            "has_more: {}",
-            result
-                .get("has_more")
-                .and_then(Value::as_bool)
-                .unwrap_or(false)
-        ),
-        format!(
-            "page_token: {}",
-            result
-                .get("page_token")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_delete_record_text(payload: &Value) -> CliResult<String> {
-    let mut lines = vec![
-        "feishu bitable delete-record".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "record_id: {}",
-            payload
-                .get("record_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "deleted: {}",
-            payload
-                .get("deleted")
-                .and_then(Value::as_bool)
-                .unwrap_or(false)
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_batch_records_text(payload: &Value) -> CliResult<String> {
-    let result = payload
-        .get("result")
-        .ok_or_else(|| "feishu bitable batch payload missing result".to_owned())?;
-    let mut lines = vec![
-        "feishu bitable batch-records".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    if let Some(records) = result.get("records").and_then(Value::as_array) {
-        lines.push(format!("records: {}", records.len()));
-    }
-    if let Some(success) = result.get("success").and_then(Value::as_bool) {
-        lines.push(format!("success: {success}"));
-    }
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_field_text(payload: &Value) -> CliResult<String> {
-    let field = payload
-        .get("field")
-        .ok_or_else(|| "feishu bitable field payload missing field".to_owned())?;
-    let mut lines = vec![
-        "feishu bitable field".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "field_id: {}",
-            field.get("field_id").and_then(Value::as_str).unwrap_or("-")
-        ),
-        format!(
-            "field_name: {}",
-            field
-                .get("field_name")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_field_list_text(payload: &Value) -> CliResult<String> {
-    let mut lines = vec![
-        "feishu bitable list-fields".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.push(format!(
-        "fields: {}",
-        payload
-            .get("fields")
-            .and_then(Value::as_array)
-            .map_or(0, std::vec::Vec::len)
-    ));
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_delete_field_text(payload: &Value) -> CliResult<String> {
-    let mut lines = vec![
-        "feishu bitable delete-field".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "field_id: {}",
-            payload
-                .get("field_id")
-                .and_then(Value::as_str)
-                .unwrap_or("-")
-        ),
-        format!(
-            "deleted: {}",
-            payload
-                .get("deleted")
-                .and_then(Value::as_bool)
-                .unwrap_or(false)
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_view_text(payload: &Value) -> CliResult<String> {
-    let view = payload
-        .get("view")
-        .ok_or_else(|| "feishu bitable view payload missing view".to_owned())?;
-    let mut lines = vec![
-        "feishu bitable view".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.extend([
-        format!(
-            "view_id: {}",
-            view.get("view_id").and_then(Value::as_str).unwrap_or("-")
-        ),
-        format!(
-            "view_name: {}",
-            view.get("view_name").and_then(Value::as_str).unwrap_or("-")
-        ),
-    ]);
-    Ok(lines.join("\n"))
-}
-
-fn render_bitable_view_list_text(payload: &Value) -> CliResult<String> {
-    let mut lines = vec![
-        "feishu bitable list-views".to_owned(),
-        format!("account: {}", required_json_string(payload, "account_id")?),
-    ];
-    if let Some(configured_account) = payload.get("configured_account").and_then(Value::as_str) {
-        lines.push(format!("configured_account: {configured_account}"));
-    }
-    lines.push(format!(
-        "views: {}",
-        payload
-            .get("views")
-            .and_then(Value::as_array)
-            .map_or(0, std::vec::Vec::len)
-    ));
-    Ok(lines.join("\n"))
-}
-
-mod render_tests {
-    use super::*;
-
-    #[allow(dead_code)]
-    fn sample_grant_summary(selected: bool, effective_selected: bool) -> Value {
-        json!({
-            "selected": selected,
-            "effective_selected": effective_selected,
-            "principal": {
-                "open_id": "ou_123",
-                "name": "Alice"
-            },
-            "status": {
-                "access_token_expired": false,
-                "refresh_token_expired": false,
-                "missing_scopes": ["docx:document:readonly"]
-            },
-            "message_write_status": {
-                "ready": false,
-                "matched_scopes": []
-            },
-            "recommendations": {
-                "auth_start_command": "loong feishu auth start --account feishu_main --capability message-write"
-            }
-        })
-    }
-
-    #[test]
-    fn render_auth_list_text_includes_stale_selection_and_select_hint() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "selected_open_id": Value::Null,
-            "grants": [sample_grant_summary(false, false)],
-            "recommendations": {
-                "select_command": "loong feishu auth select --account feishu_main --open-id <open_id>",
-                "stale_selected_open_id": "ou_missing"
-            }
-        });
-
-        let rendered = render_auth_list_text(&payload).expect("render auth list");
-
-        assert!(rendered.contains("configured_account: work"));
-        assert!(rendered.contains(
-            "select_hint: loong feishu auth select --account feishu_main --open-id <open_id>"
-        ));
-        assert!(rendered.contains("stale_selected_open_id: ou_missing"));
-        assert!(rendered.contains("missing_scopes: docx:document:readonly"));
-    }
-
-    #[test]
-    fn render_auth_select_text_includes_selected_grant_summary() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "selected_open_id": "ou_123",
-            "grant": sample_grant_summary(true, true)
-        });
-
-        let rendered = render_auth_select_text(&payload).expect("render auth select");
-
-        assert!(rendered.contains("configured_account: work"));
-        assert!(rendered.contains("selected_open_id: ou_123"));
-        assert!(
-            rendered.contains(
-                "open_id: ou_123 | selected: true | effective_selected: true | name: Alice"
-            )
-        );
-    }
-
-    #[test]
-    fn render_auth_list_text_distinguishes_persisted_and_effective_selection() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "selected_open_id": Value::Null,
-            "effective_open_id": "ou_123",
-            "grants": [sample_grant_summary(false, true)],
-            "recommendations": {}
-        });
-
-        let rendered = render_auth_list_text(&payload).expect("render auth list");
-
-        assert!(rendered.contains(
-            "open_id: ou_123 | selected: false | effective_selected: true | name: Alice"
-        ));
-    }
-
-    #[test]
-    fn render_auth_start_text_includes_configured_account_when_present() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "state": "state_123",
-            "redirect_uri": "http://127.0.0.1:34819/callback",
-            "authorize_url": "https://open.feishu.cn/open-apis/authen/v1/authorize",
-            "sqlite_path": "/tmp/feishu.sqlite3",
-            "capabilities": ["read-only"],
-            "scopes": ["offline_access", "docx:document:readonly"],
-        });
-
-        let rendered = render_auth_start_text(&payload).expect("render auth start");
-
-        assert!(rendered.contains("configured_account: work"));
-    }
-
-    #[test]
-    fn render_onboard_text_includes_qr_registration_summary() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "config": "/tmp/loong.toml",
-            "credential_source": "qr_registration",
-            "domain": "lark",
-            "mode": "websocket",
-            "owner_open_id": "ou_owner_1",
-            "bot_name": "Loong Bot",
-            "bot_open_id": "ou_bot_1",
-            "qr_url": "https://scan.example/activate",
-            "owner_direct_chat_bootstrap_applied": true,
-            "serve_command": "loong feishu serve --account work",
-            "status_command": "loong doctor",
-            "notes": ["defaulted inbound bootstrap access to `allowed_chat_ids = [\"*\"]` and `allowed_sender_ids = [\"ou_owner_1\"]` so the onboarding user can start a direct Feishu/Lark chat immediately"],
-        });
-
-        let rendered = render_onboard_text(&payload).expect("render onboard");
-
-        assert!(rendered.contains("feishu onboard"));
-        assert!(rendered.contains("configured_account: work"));
-        assert!(rendered.contains("credential_source: qr_registration"));
-        assert!(rendered.contains("bot_name: Loong Bot"));
-        assert!(rendered.contains("allowed_sender_ids = [\"ou_owner_1\"]"));
-        assert!(rendered.contains("serve_command: loong feishu serve --account work"));
-    }
-
-    #[test]
-    fn render_auth_exchange_text_includes_selected_and_effective_open_ids() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "principal": {
-                "open_id": "ou_123",
-                "name": "Alice"
-            },
-            "granted_scopes": ["offline_access", "docx:document:readonly"],
-            "selected_open_id": "ou_123",
-            "effective_open_id": "ou_123",
-        });
-
-        let rendered = render_auth_exchange_text(&payload).expect("render auth exchange");
-
-        assert!(rendered.contains("configured_account: work"));
-        assert!(rendered.contains("selected_open_id: ou_123"));
-        assert!(rendered.contains("effective_open_id: ou_123"));
-    }
-
-    #[test]
-    fn render_auth_revoke_text_includes_remaining_grant_state_and_hints() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "open_id": "ou_789",
-            "deleted": true,
-            "grant_count": 2,
-            "selected_open_id": Value::Null,
-            "effective_open_id": Value::Null,
-            "recommendations": {
-                "select_command": "loong feishu auth select --account feishu_main --open-id <open_id>"
-            }
-        });
-
-        let rendered = render_auth_revoke_text(&payload).expect("render auth revoke");
-
-        assert!(rendered.contains("configured_account: work"));
-        assert!(rendered.contains("grant_count: 2"));
-        assert!(rendered.contains("selected_open_id: -"));
-        assert!(rendered.contains("effective_open_id: -"));
-        assert!(rendered.contains(
-            "select_hint: loong feishu auth select --account feishu_main --open-id <open_id>"
-        ));
-    }
-
-    #[test]
-    fn render_auth_status_text_account_scope_includes_missing_scopes_and_select_hint() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "status_scope": "account",
-            "grant_count": 2,
-            "selected_open_id": Value::Null,
-            "grants": [sample_grant_summary(false, false)],
-            "recommendations": {
-                "select_command": "loong feishu auth select --account feishu_main --open-id <open_id>",
-                "stale_selected_open_id": "ou_missing"
-            }
-        });
-
-        let rendered = render_auth_status_text(&payload).expect("render auth status");
-
-        assert!(rendered.contains("configured_account: work"));
-        assert!(rendered.contains("status_scope: account"));
-        assert!(rendered.contains(
-            "select_hint: loong feishu auth select --account feishu_main --open-id <open_id>"
-        ));
-        assert!(rendered.contains("stale_selected_open_id: ou_missing"));
-        assert!(rendered.contains("missing_scopes: docx:document:readonly"));
-    }
-
-    #[test]
-    fn render_auth_status_text_grant_scope_includes_requested_open_id_and_available_options() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "status_scope": "grant",
-            "status": {
-                "has_grant": false,
-                "access_token_expired": false,
-                "refresh_token_expired": false,
-                "missing_scopes": []
-            },
-            "message_write_status": {
-                "ready": false,
-                "matched_scopes": []
-            },
-            "recommendations": {
-                "auth_start_command": Value::Null,
-                "select_command": "loong feishu auth select --account feishu_main --open-id <open_id>"
-            },
-            "selected_open_id": Value::Null,
-            "effective_open_id": Value::Null,
-            "requested_open_id": "ou_missing",
-            "available_open_ids": ["ou_456", "ou_123"]
-        });
-
-        let rendered = render_auth_status_text(&payload).expect("render auth status");
-
-        assert!(rendered.contains("configured_account: work"));
-        assert!(rendered.contains("requested_open_id: ou_missing"));
-        assert!(rendered.contains(
-            "select_hint: loong feishu auth select --account feishu_main --open-id <open_id>"
-        ));
-        assert!(rendered.contains("available_open_ids: ou_456, ou_123"));
-    }
-
-    #[test]
-    fn render_whoami_text_includes_configured_account_when_present() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "principal": {
-                "open_id": "ou_123",
-                "name": "Alice",
-                "email": "alice@example.com",
-                "tenant_key": "tenant_x"
-            }
-        });
-
-        let rendered = render_whoami_text(&payload).expect("render whoami");
-
-        assert!(rendered.contains("configured_account: work"));
-    }
-
-    #[test]
-    fn render_read_doc_text_includes_configured_account_when_present() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "document": {
-                "document_id": "doxcnDemo",
-                "content": "hello"
-            }
-        });
-
-        let rendered = render_read_doc_text(&payload).expect("render read doc");
-
-        assert!(rendered.contains("configured_account: work"));
-    }
-
-    #[test]
-    fn render_doc_create_text_includes_configured_account_when_present() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "document": {
-                "document_id": "doxcnCreated",
-                "url": "https://open.feishu.cn/docx/doxcnCreated"
-            },
-            "content_inserted": true,
-            "inserted_block_count": 1,
-            "insert_batch_count": 1,
-            "content_type": "markdown"
-        });
-
-        let rendered = render_doc_create_text(&payload).expect("render doc create");
-
-        assert!(rendered.contains("configured_account: work"));
-        assert!(rendered.contains("content_inserted: true"));
-        assert!(rendered.contains("insert_batch_count: 1"));
-    }
-
-    #[test]
-    fn render_doc_append_text_includes_configured_account_when_present() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "document": {
-                "document_id": "doxcnExisting",
-                "url": "https://open.feishu.cn/docx/doxcnExisting"
-            },
-            "inserted_block_count": 1,
-            "insert_batch_count": 1,
-            "content_type": "markdown"
-        });
-
-        let rendered = render_doc_append_text(&payload).expect("render doc append");
-
-        assert!(rendered.contains("configured_account: work"));
-        assert!(rendered.contains("inserted_block_count: 1"));
-        assert!(rendered.contains("insert_batch_count: 1"));
-    }
-
-    #[test]
-    fn render_messages_history_text_includes_configured_account_when_present() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "page": {
-                "items": [{"message_id": "om_1"}],
-                "has_more": true,
-                "page_token": "next-page"
-            }
-        });
-
-        let rendered = render_messages_history_text(&payload).expect("render messages history");
-
-        assert!(rendered.contains("configured_account: work"));
-    }
-
-    #[test]
-    fn render_messages_get_text_includes_configured_account_when_present() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "message": {
-                "message_id": "om_1"
-            }
-        });
-
-        let rendered = render_messages_get_text(&payload).expect("render messages get");
-
-        assert!(rendered.contains("configured_account: work"));
-    }
-
-    #[test]
-    fn render_messages_resource_text_includes_configured_account_when_present() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "message_id": "om_resource_1",
-            "file_key": "file_resource_1",
-            "resource_type": "file",
-            "path": "/tmp/spec-sheet.pdf",
-            "bytes_written": 18,
-            "content_type": "application/pdf",
-            "file_name": "spec-sheet.pdf"
-        });
-
-        let rendered = render_messages_resource_text(&payload).expect("render messages resource");
-
-        assert!(rendered.contains("configured_account: work"));
-    }
-
-    #[test]
-    fn render_send_text_includes_configured_account_when_present() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "delivery": {
-                "message_id": "om_1",
-                "receive_id_type": "chat_id",
-                "receive_id": "oc_1",
-                "uuid": "uuid-1",
-                "msg_type": "text"
-            }
-        });
-
-        let rendered = render_send_text(&payload).expect("render send");
-
-        assert!(rendered.contains("configured_account: work"));
-    }
-
-    #[test]
-    fn render_reply_text_includes_configured_account_when_present() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "delivery": {
-                "message_id": "om_2",
-                "reply_to_message_id": "om_1",
-                "reply_in_thread": true,
-                "uuid": "uuid-2",
-                "msg_type": "interactive"
-            }
-        });
-
-        let rendered = render_reply_text(&payload).expect("render reply");
-
-        assert!(rendered.contains("configured_account: work"));
-    }
-
-    #[test]
-    fn render_search_messages_text_includes_configured_account_when_present() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "page": {
-                "items": [{"message_id": "om_1"}],
-                "has_more": false,
-                "page_token": "page-1"
-            }
-        });
-
-        let rendered = render_search_messages_text(&payload).expect("render search messages");
-
-        assert!(rendered.contains("configured_account: work"));
-    }
-
-    #[test]
-    fn render_calendar_list_text_includes_configured_account_when_present() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "primary": true,
-            "calendars": {
-                "calendars": [{
-                    "calendar": {
-                        "calendar_id": "cal_1"
-                    }
-                }]
-            }
-        });
-
-        let rendered = render_calendar_list_text(&payload).expect("render calendar list");
-
-        assert!(rendered.contains("configured_account: work"));
-    }
-
-    #[test]
-    fn render_calendar_freebusy_text_includes_configured_account_when_present() {
-        let payload = json!({
-            "account_id": "feishu_main",
-            "configured_account": "work",
-            "result": {
-                "freebusy_list": [{
-                    "start_time": "2026-03-12T09:00:00+08:00",
-                    "end_time": "2026-03-12T10:00:00+08:00"
-                }]
-            }
-        });
-
-        let rendered = render_calendar_freebusy_text(&payload).expect("render calendar freebusy");
-
-        assert!(rendered.contains("configured_account: work"));
-    }
 }
 
 #[cfg(test)]
